requirements.txt
azure-core==1.30.2
azure-functions==1.20.0
azure-storage-blob==12.22.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
idna==3.8
isodate==0.6.1
pillow==10.4.0
pycparser==2.22
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
